Change output interface to the new JsonRcpOutputStream interface
JsonRcpOutputStream is an abstract class with a single "send"
method. It should be subclassed to send data over a socket or
equivalent.

This replaces JsonRpcStream as the type of the 'output' parameter
given to JsonRpc's constructor. There's not much functional
difference, just a little easier to read and fewer methods to
override.

This change requires some minor modifications in conveyor.