JsonRpc interface cleanup: use a weak_ptr for client methods
The addMethod() member now takes a weak_ptr to a JsonRpcMethod rather
than a direct pointer. This allows clients to remove a method by
simply destroying all shared_ptr references to the JsonRpcMethod.

Use caution on the client side: if the method subclass has internal
state that may become invalidated, additional locking may be needed to
verify that no dangling references are used. The weak_ptr interface
does not protect against this sort of error.

Note that this change to the pulic API requires some minor fixes in
conveyor-cpp.