Add a way to use shared_ptrs for method callbacks
MW-5636
http://jira.makerbot.net/browse/MW-5636
We want to be able to use std::future to receive rpc responses, but this
leaves only the jsonrpc library with a reference to the callback.  I think
it is cleaner to add a way for jsonrpc to actually own a callback reference
than to force the std::future wrapping code to keep a set of callback
references for all unfulfilled promises.

The shared_ptr based invoke is a separate function since most existing code
does not explicitly convert to a weak_ptr.  I do not expect that we will
want to migrate all of our code away from using weak_ptr, so we will need
to support both methods in a way that is maintainable.