Remove trailing newlines from sent json packets
There is an option not to put the newline in in the first place, but only in
a newer release of jsoncpp than what ships on 14.04.  We could also use a
writer that is not Json::FastWriter, but I think that stripping the newline
is a better option.  (We need to strip the newline because jsonrpc raw mode
expects a binary packet to immediately follow the closing }.)