Output stream API cleanup
MW-5874
http://jira.makerbot.net/browse/MW-5874
There is no point in requiring every client that uses jsonrpc to implement
multiple overloads of send for the output stream.  Specifying the buffer
type as const char * const is also pretty silly since it is not really our
business to tell the client what is can and cannot do with its own private
copy of a pointer.